+++ crates/flashq-http/Cargo.toml
@@ -33,3 +33,4 @@ serde_yaml = "0.9"
 uuid = { version = "1.18.0", features = ["v4"] }
 tempfile = "3.13"
 tokio = { version = "1.47.1", features = ["full"] }
+env_logger.workspace = true
