+++ crates/flashq/Cargo.toml
@@ -5,11 +5,11 @@ edition.workspace = true
 license.workspace = true
 
 [lib]
-path = "../../src/lib.rs"
+path = "src/lib.rs"
 
 [[bin]]
 name = "flashq"
-path = "../../src/main.rs"
+path = "src/main.rs"
 
 [dependencies]
 serde.workspace = true
@@ -26,6 +26,7 @@ serde_yaml = "0.9"
 uuid = { version = "1.18.0", features = ["v4"] }
 tempfile = "3.13"
 divan = "0.1"
+env_logger.workspace = true
 
 [[bench]]
 name = "memory_storage"
