+++ crates/flashq/src/error.rs
@@ -251,4 +251,3 @@ mod tests {
         }
     }
 }
-
