+++ crates/flashq/src/main.rs
@@ -1,4 +1,3 @@
 fn main() {
     flashq::demo::run_interactive_demo();
 }
-
