--- src/main.rs
@@ -1,3 +1,0 @@
-fn main() {
-    flashq::demo::run_interactive_demo();
-}
