--- src/storage/file/mod.rs
@@ -1,16 +1,0 @@
-pub mod common;
-pub mod consumer_group;
-pub mod file_io;
-pub mod index;
-pub mod segment;
-pub mod segment_manager;
-pub mod time_index;
-pub mod topic_log;
-
-pub use common::SyncMode;
-pub use consumer_group::FileConsumerGroup;
-pub use file_io::FileIo;
-pub use index::{IndexEntry, SparseIndex};
-pub use segment::{IndexingConfig, LogSegment};
-pub use segment_manager::SegmentManager;
-pub use topic_log::FileTopicLog;
