--- src/storage/mod.rs
@@ -1,10 +1,0 @@
-pub mod backend;
-pub mod batching_heuristics;
-pub mod file;
-pub mod memory;
-pub mod r#trait;
-
-// Re-exports for ergonomics
-pub use backend::StorageBackend;
-pub use memory::{InMemoryConsumerGroup, InMemoryTopicLog};
-pub use r#trait::{ConsumerGroup, TopicLog};
